requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
scikit-learn==1.5.0
numpy==1.26.4
pandas==2.2.2
joblib==1.4.2
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
